requirements.txt
attrs==23.2.0
azure-ai-ml==1.18.0
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.17.1
azure-mgmt-core==1.4.0
azure-storage-blob==12.21.0
azure-storage-file-datalake==12.16.0
azure-storage-file-share==12.17.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.0
distlib==0.3.8
exceptiongroup==1.2.2
filelock==3.15.4
google-api-core==2.19.1
google-auth==2.32.0
googleapis-common-protos==1.63.2
identify==2.6.0
idna==3.7
iniconfig==2.0.0
isodate==0.6.1
jsonargparse==4.32.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
loguru==0.7.2
marshmallow==3.21.3
mldesigner==0.1.0b18
msal==1.30.0
msal-extensions==1.2.0
msrest==0.7.1
nodeenv==1.9.1
oauthlib==3.2.2
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.13
opencensus-ext-logging==0.1.1
packaging==24.1
platformdirs==4.2.2
pluggy==1.5.0
portalocker==2.10.1
pre-commit==3.8.0
proto-plus==1.24.0
protobuf==5.27.3
psutil==6.0.0
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
pydash==8.0.3
pyjwt==2.9.0
pytest==8.3.2
python-dateutil==2.9.0.post0
pyyaml==6.0.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.19.1
rsa==4.9
six==1.16.0
strictyaml==1.7.3
tomli==2.0.1
tqdm==4.66.4
typing-extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.3
